+++ router/cmd/instance.go
@@ -46,18 +46,18 @@ func NewRouter(params Params, additionalOptions ...core.Option) (*core.Router, e
 			logger.Fatal("Could not read router config", zap.Error(err), zap.String("path", cfg.RouterConfigPath))
 		}
 	} else {
-		routerCDN, err := cdn.New(cdn.CDNOptions{
-			URL:                 cfg.CDN.URL,
-			AuthenticationToken: cfg.Graph.Token,
-			CacheSize:           cfg.CDN.CacheSize.Uint64(),
+		routerCDN, err := cdn.NewRouterConfigClient(cfg.CDN.URL, cfg.Graph.Token, cdn.PersistentOperationsOptions{
+			CacheSize: cfg.CDN.CacheSize.Uint64(),
+			Logger:    logger,
 		})
 		if err != nil {
 			return nil, err
 		}
 
-		configPoller = configpoller.New(cfg.Graph.Name, cfg.ControlplaneURL, cfg.Graph.Token, routerCDN,
+		configPoller = configpoller.New(cfg.Graph.Name, cfg.ControlplaneURL, cfg.Graph.Token,
 			configpoller.WithLogger(logger),
 			configpoller.WithPollInterval(cfg.PollInterval),
+			configpoller.WithCDNClient(routerCDN),
 		)
 	}
