--- router/internal/cdn/doc.go
@@ -1,2 +1,0 @@
-// Package cdn implements primitives to talk to the Cosmo CDN
-package cdn
