+++ router/internal/controlplane/configpoller/config_poller.go
@@ -24,6 +24,7 @@ type ConfigPoller interface {
 	// to execute, the next invocation will be skipped.
 	Subscribe(ctx context.Context, handler func(newConfig *nodev1.RouterConfig, oldVersion string) error)
 	// GetRouterConfig returns the latest router config from the controlplane
+	// If the Config is nil, no new config is available and the current config should be used.
 	// and updates the latest router config version. This method is only used for the initial config
 	GetRouterConfig(ctx context.Context) (*nodev1.RouterConfig, error)
 	// Stop stops the config poller. After calling stop, the config poller cannot be used again.
@@ -39,15 +40,14 @@ type configPoller struct {
 	latestRouterConfigVersion string
 	poller                    controlplane.Poller
 	pollInterval              time.Duration
-	cdn                       *cdn.CDN
+	cdnConfigClient           *cdn.RouterConfigClient
 }
 
-func New(graphName, endpoint, token string, cdn *cdn.CDN, opts ...Option) ConfigPoller {
+func New(graphName, endpoint, token string, opts ...Option) ConfigPoller {
 	c := &configPoller{
 		controlplaneEndpoint: endpoint,
 		graphApiToken:        token,
 		federatedGraphName:   graphName,
-		cdn:                  cdn,
 	}
 
 	for _, opt := range opts {
@@ -65,7 +65,7 @@ func New(graphName, endpoint, token string, cdn *cdn.CDN, opts ...Option) Config
 	retryClient.Logger = nil
 	retryClient.RequestLogHook = func(_ retryablehttp.Logger, _ *http.Request, retry int) {
 		if retry > 0 {
-			c.logger.Info("Fetch router config", zap.Int("retry", retry))
+			c.logger.Info("Fetch router config from controlplane", zap.Int("retry", retry))
 		}
 	}
 
@@ -93,24 +93,14 @@ func (c *configPoller) Stop(_ context.Context) error {
 func (c *configPoller) Subscribe(ctx context.Context, handler func(newConfig *nodev1.RouterConfig, oldVersions string) error) {
 
 	c.poller.Subscribe(ctx, func() {
-		cfg, err := c.cdn.RouterConfig(ctx, c.latestRouterConfigVersion)
+		cfg, err := c.getRouterConfig(ctx)
 		// both being nil indicates that no new router config available
 		if cfg == nil && err == nil {
-			c.logger.Debug("No new router config available, received nil router config, trying again in 10 seconds")
+			c.logger.Sugar().Debugf("No new router config available. Trying again in %s", c.pollInterval.String())
 			return
 		}
-		if err != nil {
-			cfg, err := c.getRouterConfigFromCP(ctx, &c.latestRouterConfigVersion)
-			if err != nil {
-				return
-			}
-			if cfg == nil {
-				c.logger.Debug("No new router config available, received nil router config, trying again in 10 seconds")
-				return
-			}
-		}
-		newVersion := cfg.GetVersion()
 
+		newVersion := cfg.GetVersion()
 		latestVersion := c.latestRouterConfigVersion
 
 		// If the version hasn't changed, don't invoke the handler
@@ -161,32 +151,33 @@ func (c *configPoller) getRouterConfigFromCP(ctx context.Context, version *strin
 	return resp.Msg.GetConfig(), nil
 }
 
-// GetRouterConfig returns the latest router config from the cdn first, if not found then it fetches from the controlplane. Not safe for concurrent use.
-func (c *configPoller) GetRouterConfig(ctx context.Context) (*nodev1.RouterConfig, error) {
-	c.logger.Info("Fetching initial router configuration from cdn")
-
-	cfg, err := c.cdn.RouterConfig(ctx, c.latestRouterConfigVersion)
-	if cfg != nil {
-		c.latestRouterConfigVersion = cfg.GetVersion()
+func (c *configPoller) getRouterConfig(ctx context.Context) (*nodev1.RouterConfig, error) {
+	cfg, err := c.cdnConfigClient.RouterConfig(ctx, c.latestRouterConfigVersion)
+	if err == nil {
 		return cfg, nil
 	}
 
-	c.logger.Warn("Failed to fetch the router config from cdn.", zap.Error(err))
-
-	c.logger.Info("Fetching initial router configuration from control plane")
+	c.logger.Warn("Fallback fetching initial router configuration from control plane")
 
 	cfg, err = c.getRouterConfigFromCP(ctx, nil)
+	if err == nil {
+		return cfg, nil
+	}
+
 	if err != nil {
 		return nil, err
 	}
 
-	if cfg == nil {
-		return nil, fmt.Errorf("received nil router config from control plane")
-	}
-
-	c.latestRouterConfigVersion = cfg.GetVersion()
-
 	return cfg, nil
+}
+
+// GetRouterConfig returns the latest router config from the cdnConfigClient first, if not found then it fetches from the controlplane. Not safe for concurrent use.
+func (c *configPoller) GetRouterConfig(ctx context.Context) (*nodev1.RouterConfig, error) {
+	cfg, err := c.getRouterConfig(ctx)
+	if err == nil {
+		c.latestRouterConfigVersion = cfg.GetVersion()
+	}
+	return cfg, err
 }
 
 func WithLogger(logger *zap.Logger) Option {
@@ -198,5 +189,11 @@ func WithLogger(logger *zap.Logger) Option {
 func WithPollInterval(interval time.Duration) Option {
 	return func(s *configPoller) {
 		s.pollInterval = interval
+	}
+}
+
+func WithCDNClient(cdnConfigClient *cdn.RouterConfigClient) Option {
+	return func(s *configPoller) {
+		s.cdnConfigClient = cdnConfigClient
 	}
 }
